pyproject.toml
[project]
name = "grantflow"
version = "0.1.0"
description = "Unified government grants & contracts data API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "python-dotenv>=1.2.0",
    "httpx>=0.27",
    "lxml>=5.0",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",
    "structlog>=24.0",
    "apscheduler>=3.11,<4.0",
    "slowapi>=0.1.9",
    "scrapling>=0.4.2",
    "curl-cffi>=0.14.0",
    "playwright>=1.58.0",
    "browserforge>=1.2.4",
    "instructor>=1.14.5",
    "openai>=2.29.0",
]

[project.scripts]
grantflow = "grantflow.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["grantflow*"]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24",
]
